requirements.txt
python-dotenv==1.0.0
vk-api==11.9.9